\name{init.data}
\docType{data}
\alias{init.data}
\title{Title string for init.data}
\description{
   This is the description of the init.data dataset.
}
\usage{init.data}
\format{This describes the format of init.data.}
\source{This gives the source of init.data}
\references{
   Any references for init.data go here
}
